+++ code/game-server-v2/GameSettings.ts
