+++ code/game-server-v2/testing/client.ts
